+++ multisrc/overrides/madara/ichirinnohanayuri/src/IchirinNoHanaYuri.kt
@@ -1,0 +1,7 @@
+package eu.kanade.tachiyomi.extension.en.ichirinnohanayuri
+
+import eu.kanade.tachiyomi.multisrc.madara.Madara
+import java.text.SimpleDateFormat
+import java.util.Locale
+
+class IchirinNoHanaYuri : Madara("Ichirin No Hana Yuri", "https://ichirinnohanayuri.com.br", "pt-BR", SimpleDateFormat("dd/MM/yyyy", Locale("pt")))
