+++ multisrc/overrides/madara/kombatch/src/Kombatch.kt
@@ -1,0 +1,7 @@
+package eu.kanade.tachiyomi.extension.en.kombatch
+
+import eu.kanade.tachiyomi.multisrc.madara.Madara
+import java.text.SimpleDateFormat
+import java.util.Locale
+
+class Kombatch : Madara("Kombatch", "https://kombatch.com", "id", SimpleDateFormat("d MMMM yyyy", Locale.forLanguageTag("id")))
